requirements.txt
allure-pytest==2.8.24
allure-python-commons==2.8.24
attrs==20.3.0
beautifulsoup4==4.9.1
certifi==2020.4.5.1
chardet==3.0.4
contextlib2==0.6.0.post1
et-xmlfile==1.0.1
future==0.18.2
gevent==20.9.0
greenlet==0.4.17
idna==2.9
iniconfig==1.1.1
itchat==1.2.32
jdcal==1.4.1
lxml==4.6.2
numpy==1.20.0
openpyxl==3.0.6
packaging==20.8
pandas==1.2.0
pluggy==0.13.1
py==1.10.0
pycryptodome==3.9.8
pyparsing==2.4.7
pypng==0.0.20
PyQRCode==1.2.1
pytest==6.2.1
pytest-html==3.1.1
pytest-metadata==1.11.0
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.3.1
requests==2.24.0
schema==0.7.4
selenium==3.141.0
six==1.15.0
soupsieve==2.1
toml==0.10.2
urllib3==1.25.9
websocket-client==0.57.0
websockets==8.1
wxpy==0.3.9.8
xlrd==2.0.1
xlwt==1.3.0
zope.event==4.5.0
zope.interface==5.2.0
requirements.txt